added wrapper class around PLCamera + major restructure
