+++ src/engine/editor/window_map_config.cpp
@@ -22,7 +22,7 @@
 using namespace openhow;
 
 MapConfigEditor::MapConfigEditor() {
-  map_ = engine->GetGameManager()->GetCurrentMap();
+  map_ = Engine::GameManagerInstance()->GetCurrentMap();
   if (map_ == nullptr) {
     throw std::runtime_error("Attempted to create config editor without a valid map loaded!\n");
   }
