+++ src/engine/editor/window_new_map.cpp
@@ -28,8 +28,8 @@ void NewMapWindow::Display() {
   //ImGui::InputText("Author", author_buffer_, sizeof(author_buffer_));
   ImGui::Separator();
   if(ImGui::Button("Create")) {
-    engine->GetGameManager()->CreateManifest(name_buffer_);
-    engine->GetGameManager()->LoadMap(name_buffer_);
+    Engine::GameManagerInstance()->CreateManifest(name_buffer_);
+    Engine::GameManagerInstance()->LoadMap(name_buffer_);
     SetStatus(false);
   }
   ImGui::SameLine();
