+++ src/engine/engine.h
@@ -53,10 +53,19 @@
 #include "game/game.h"
 
 namespace openhow {
+class Engine;
+extern Engine* engine;
 class Engine {
  public:
   Engine();
   ~Engine();
+
+  static AudioManager* AudioManagerInstance() {
+    return engine->audio_manager_;
+  }
+  static GameManager* GameManagerInstance() {
+    return engine->game_manager_;
+  }
 
   void Initialize();
 
@@ -64,16 +73,10 @@ class Engine {
 
   bool IsRunning();
 
-  // Subsystems
-  GameManager* GetGameManager() { return game_manager_; }
-  AudioManager* GetAudioManager() { return audio_manager_; }
-
  private:
   GameManager* game_manager_{nullptr};
   AudioManager* audio_manager_{nullptr};
 };
-
-extern Engine* engine;
 }
 
 #else
@@ -81,7 +84,6 @@ typedef struct BaseGameMode BaseGameMode;
 #endif // __cplusplus; todo: remove this once all code is compiled as C++
 
 typedef struct EngineState {
-  struct PLCamera* camera;       // camera used for general gameplay
   struct PLCamera* ui_camera;    // camera used for UI elements, orthographic
 
   unsigned int sys_ticks;
