+++ src/engine/game/game.cpp
@@ -81,6 +81,8 @@ GameManager::GameManager() {
   plRegisterConsoleCommand("createmap", CreateMapCommand, "");
   plRegisterConsoleCommand("map", MapCommand, "");
   plRegisterConsoleCommand("maps", MapsCommand, "");
+
+  camera_ = new Camera({0, 0, 0}, {0, 0, 0});
 }
 
 GameManager::~GameManager() {
@@ -102,7 +104,7 @@ void GameManager::Tick() {
           active_map_->GetTerrain()->GetMaxHeight(),
           plGenerateRandomf(TERRAIN_PIXEL_WIDTH)
       };
-      engine->GetAudioManager()->PlayLocalSound(sample, position, {0, 0, 0}, true, 0.5f);
+      Engine::AudioManagerInstance()->PlayLocalSound(sample, position, {0, 0, 0}, true, 0.5f);
     }
 
     ambient_emit_delay_ = g_state.sim_ticks + TICKS_PER_SECOND + rand() % (7 * TICKS_PER_SECOND);
@@ -112,7 +114,7 @@ void GameManager::Tick() {
 }
 
 void GameManager::LoadMap(const std::string& name) {
-  MapManifest* manifest = engine->GetGameManager()->GetMapManifest(name);
+  MapManifest* manifest = Engine::GameManagerInstance()->GetMapManifest(name);
   if (manifest == nullptr) {
     LogWarn("Failed to get map descriptor, \"%s\"\n", name.c_str());
     return;
@@ -136,13 +138,13 @@ void GameManager::LoadMap(const std::string& name) {
 
   for (unsigned int i = 1, idx = 0; i < 4; ++i) {
     if (i < 3) {
-      ambient_samples_[idx++] = engine->GetAudioManager()->CacheSample(
+      ambient_samples_[idx++] = Engine::AudioManagerInstance()->CacheSample(
           "audio/amb_" + std::to_string(i) + sample_ext + ".wav", false);
     }
     ambient_samples_[idx++] =
-        engine->GetAudioManager()->CacheSample("audio/batt_s" + std::to_string(i) + ".wav", false);
+        Engine::AudioManagerInstance()->CacheSample("audio/batt_s" + std::to_string(i) + ".wav", false);
     ambient_samples_[idx++] =
-        engine->GetAudioManager()->CacheSample("audio/batt_l" + std::to_string(i) + ".wav", false);
+        Engine::AudioManagerInstance()->CacheSample("audio/batt_l" + std::to_string(i) + ".wav", false);
   }
 
   ambient_emit_delay_ = g_state.sim_ticks + rand() % 100;
@@ -204,7 +206,7 @@ void GameManager::RegisterMapManifest(const std::string& path) {
 }
 
 static void RegisterManifestInterface(const char* path) {
-  engine->GetGameManager()->RegisterMapManifest(path);
+  Engine::GameManagerInstance()->RegisterMapManifest(path);
 }
 
 void GameManager::RegisterMapManifests() {
@@ -226,7 +228,7 @@ MapManifest* GameManager::GetMapManifest(const std::string& name) {
 
 MapManifest* GameManager::CreateManifest(const std::string& name) {
   // ensure the map doesn't exist already
-  if(engine->GetGameManager()->GetMapManifest(name) != nullptr) {
+  if(Engine::GameManagerInstance()->GetMapManifest(name) != nullptr) {
     LogWarn("Unable to create map, it already exists!\n");
     return nullptr;
   }
@@ -244,8 +246,8 @@ MapManifest* GameManager::CreateManifest(const std::string& name) {
 
   LogInfo("Wrote \"%s\"!\n", path.c_str());
 
-  engine->GetGameManager()->RegisterMapManifest(path);
-  return engine->GetGameManager()->GetMapManifest(name);
+  Engine::GameManagerInstance()->RegisterMapManifest(path);
+  return Engine::GameManagerInstance()->GetMapManifest(name);
 }
 
 void GameManager::CreateMapCommand(unsigned int argc, char** argv) {
@@ -254,12 +256,12 @@ void GameManager::CreateMapCommand(unsigned int argc, char** argv) {
     return;
   }
 
-  MapManifest* manifest = engine->GetGameManager()->CreateManifest(argv[1]);
+  MapManifest* manifest = Engine::GameManagerInstance()->CreateManifest(argv[1]);
   if(manifest == nullptr) {
     return;
   }
 
-  engine->GetGameManager()->LoadMap(argv[1]);
+  Engine::GameManagerInstance()->LoadMap(argv[1]);
 }
 
 void GameManager::MapCommand(unsigned int argc, char** argv) {
@@ -269,21 +271,21 @@ void GameManager::MapCommand(unsigned int argc, char** argv) {
   }
 
   std::string mode = "singleplayer";
-  const MapManifest* desc = engine->GetGameManager()->GetMapManifest(argv[1]);
+  const MapManifest* desc = Engine::GameManagerInstance()->GetMapManifest(argv[1]);
   if (desc != nullptr && !desc->modes.empty()) {
     mode = desc->modes[0];
   }
 
-  engine->GetGameManager()->LoadMap(argv[1]);
+  Engine::GameManagerInstance()->LoadMap(argv[1]);
 }
 
 void GameManager::MapsCommand(unsigned int argc, char** argv) {
-  if (engine->GetGameManager()->map_manifests_.empty()) {
+  if (Engine::GameManagerInstance()->map_manifests_.empty()) {
     LogWarn("No maps available!\n");
     return;
   }
 
-  for (auto manifest : engine->GetGameManager()->map_manifests_) {
+  for (auto manifest : Engine::GameManagerInstance()->map_manifests_) {
     MapManifest* desc = &manifest.second;
     std::string out =
         desc->name + "/" + manifest.first +
@@ -296,5 +298,5 @@ void GameManager::MapsCommand(unsigned int argc, char** argv) {
     LogInfo("%s\n", out.c_str());
   }
 
-  LogInfo("%u maps\n", engine->GetGameManager()->map_manifests_.size());
+  LogInfo("%u maps\n", Engine::GameManagerInstance()->map_manifests_.size());
 }
